src/Model_Training/requirements.txt
tensorflow
sklearn
pandas
pyyaml
tf2onnx
seaborn
matplotlib
tensorflow-addons
dvclive
optuna
plotly
